setup the basic endpoints for table
