+++ .idea/libraries/Maven__org_mongodb_bson_4_6_1.xml
@@ -1,0 +1,13 @@
+<component name="libraryTable">
+  <library name="Maven: org.mongodb:bson:4.6.1">
+    <CLASSES>
+      <root url="jar://$MAVEN_REPOSITORY$/org/mongodb/bson/4.6.1/bson-4.6.1.jar!/" />
+    </CLASSES>
+    <JAVADOC>
+      <root url="jar://$MAVEN_REPOSITORY$/org/mongodb/bson/4.6.1/bson-4.6.1-javadoc.jar!/" />
+    </JAVADOC>
+    <SOURCES>
+      <root url="jar://$MAVEN_REPOSITORY$/org/mongodb/bson/4.6.1/bson-4.6.1-sources.jar!/" />
+    </SOURCES>
+  </library>
+</component>
