+++ .idea/libraries/Maven__org_springframework_boot_spring_boot_starter_data_mongodb_2_7_3.xml
@@ -1,0 +1,13 @@
+<component name="libraryTable">
+  <library name="Maven: org.springframework.boot:spring-boot-starter-data-mongodb:2.7.3">
+    <CLASSES>
+      <root url="jar://$MAVEN_REPOSITORY$/org/springframework/boot/spring-boot-starter-data-mongodb/2.7.3/spring-boot-starter-data-mongodb-2.7.3.jar!/" />
+    </CLASSES>
+    <JAVADOC>
+      <root url="jar://$MAVEN_REPOSITORY$/org/springframework/boot/spring-boot-starter-data-mongodb/2.7.3/spring-boot-starter-data-mongodb-2.7.3-javadoc.jar!/" />
+    </JAVADOC>
+    <SOURCES>
+      <root url="jar://$MAVEN_REPOSITORY$/org/springframework/boot/spring-boot-starter-data-mongodb/2.7.3/spring-boot-starter-data-mongodb-2.7.3-sources.jar!/" />
+    </SOURCES>
+  </library>
+</component>
