+++ .idea/libraries/Maven__org_springframework_data_spring_data_mongodb_3_4_2.xml
@@ -1,0 +1,13 @@
+<component name="libraryTable">
+  <library name="Maven: org.springframework.data:spring-data-mongodb:3.4.2">
+    <CLASSES>
+      <root url="jar://$MAVEN_REPOSITORY$/org/springframework/data/spring-data-mongodb/3.4.2/spring-data-mongodb-3.4.2.jar!/" />
+    </CLASSES>
+    <JAVADOC>
+      <root url="jar://$MAVEN_REPOSITORY$/org/springframework/data/spring-data-mongodb/3.4.2/spring-data-mongodb-3.4.2-javadoc.jar!/" />
+    </JAVADOC>
+    <SOURCES>
+      <root url="jar://$MAVEN_REPOSITORY$/org/springframework/data/spring-data-mongodb/3.4.2/spring-data-mongodb-3.4.2-sources.jar!/" />
+    </SOURCES>
+  </library>
+</component>
