+++ src/main/java/com/mdma/tableservice/model/Table.java
@@ -1,5 +1,6 @@
-package com.mdma.tableservice;
+package com.mdma.tableservice.model;
 
+import lombok.AllArgsConstructor;
 import org.springframework.core.StandardReflectionParameterNameDiscoverer;
 import org.springframework.data.annotation.Id;
 import org.springframework.data.mongodb.core.mapping.Document;
@@ -9,17 +10,13 @@ import java.util.UUID;
 
 @Data
 @Document
+@AllArgsConstructor
 public class Table {
     @Id
     private String id;
 
+    private String restaurantId;
     private int tableNumber;
     private List<String> orders;
     private String reservationTime;
-
-    public Table(int tableNumber, List<String> orders, String reservationTime) {
-        this.tableNumber = tableNumber;
-        this.orders = orders;
-        this.reservationTime = reservationTime;
-    }
 }
